Refactor Mongo helper and KYC API
