+++ .env.example
@@ -3,5 +3,5 @@ MONGO_INITDB_ROOT_USERNAME=
 MONGO_INITDB_ROOT_PASSWORD=
 
 # Application database connection
-MONGO_URI=
-MONGO_DB_NAME=
+MONGODB_URI=
+MONGODB_DB=
