--- lib/mongodb-client.js
@@ -1,25 +1,0 @@
-import { MongoClient } from "mongodb";
-
-const uri = process.env.MONGO_URI;
-const options = {};
-
-let client;
-let clientPromise;
-
-if (!uri) {
-  // Return a rejected promise instead of throwing so API routes can handle it
-  clientPromise = Promise.reject(
-    new Error("Please define the MONGO_URI environment variable inside .env.local"),
-  );
-} else if (process.env.NODE_ENV === "development") {
-  if (!global._mongoClientPromise) {
-    client = new MongoClient(uri, options);
-    global._mongoClientPromise = client.connect();
-  }
-  clientPromise = global._mongoClientPromise;
-} else {
-  client = new MongoClient(uri, options);
-  clientPromise = client.connect();
-}
-
-export default clientPromise;
