+++ pages/api/kyc.js
@@ -1,20 +1,25 @@
-import clientPromise from "../../lib/mongodb-client";
+import clientPromise from "../../lib/mongo";
+
+export const config = {
+  api: {
+    bodyParser: { sizeLimit: "10mb" },
+  },
+};
 
 export default async function handler(req, res) {
-  if (req.method !== "POST") {
-    res.setHeader("Allow", ["POST"]);
-    return res.status(405).end(`Method ${req.method} Not Allowed`);
-  }
-
   try {
-    const client = await clientPromise;
-    const db = client.db(process.env.MONGO_DB_NAME || "kycdb");
-    const { fullName, email, country, documentType, documentNumber } = req.body;
-
-    if (!fullName || !email || !country || !documentType || !documentNumber) {
-      return res.status(400).json({ message: "Missing required fields" });
+    if (req.method !== "POST") {
+      res.setHeader("Allow", ["POST"]);
+      return res.status(405).json({ message: "Method not allowed" });
     }
 
+    const { fullName, email, country, documentType, documentNumber } = req.body || {};
+    if (!fullName || !email || !country || !documentType || !documentNumber) {
+      return res.status(400).json({ message: "Invalid payload" });
+    }
+
+    const client = await clientPromise;
+    const db = client.db(process.env.MONGODB_DB || "kycdb");
     const result = await db.collection("kyc").insertOne({
       fullName,
       email,
@@ -24,12 +29,9 @@ export default async function handler(req, res) {
       createdAt: new Date(),
     });
 
-    return res.status(201).json({ insertedId: result.insertedId });
-  } catch (error) {
-    console.error("Error inserting KYC data", error);
-    const message = error.message?.includes("MONGO_URI")
-      ? "Database connection not configured"
-      : "Failed to save KYC data";
-    return res.status(500).json({ message });
+    return res.status(200).json({ ok: true, id: String(result.insertedId) });
+  } catch (e) {
+    console.error("KYC API error:", e);
+    return res.status(500).json({ message: "Failed to save KYC data" });
   }
 }
